+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## X.X.X
+* Added a call to enroll users to A/B tests when getting a remote config value: 'getValueAndEnroll'
+* Added a call to enroll users to A/B tests when getting all remote config values: 'getAllValuesAndEnroll'
+
 * Mitigated an issue where users could not enroll to an A/B tests if enrollment request has failed
 
 ## 23.8.2